feat(desktop): add glow keyframes and refactor AISummary import
- Introduced a new glow keyframe animation to enhance visual effects in the application.
- Refactored the AISummary component import in both mobile and desktop entry content files for improved code organization.

These changes aim to elevate the user experience through enhanced animations and cleaner code structure.
